+++ software/controller/lib/debug/interface.h
@@ -77,7 +77,7 @@ class Interface {
 
   // Buffer into which request data is written in AwaitingCommand state
   uint8_t request_[500] = {0};
-  uint32_t request_size_{0};
+  size_t request_size_{0};
   // Remember when we receive an escape char (in case the call happens in
   // between the escape char and the special char)
   bool escape_next_byte_{false};
@@ -85,14 +85,14 @@ class Interface {
   // Buffer into which the command handler writes its response and which is then
   // sent in Responding state.
   uint8_t response_[500] = {0};
-  uint32_t response_size_{0};
-  uint32_t response_bytes_sent_{0};
+  size_t response_size_{0};
+  size_t response_bytes_sent_{0};
 
   // Some commands take time to be fully processed, we record their start time
   // and status
   Time command_start_time_{microsSinceStartup(0)};
   bool command_processed_{true};
-  uint32_t response_length_{0};
+  size_t response_length_{0};
 
   // List of registered command handlers
   Command::Handler *registry_[32] = {nullptr};
@@ -107,7 +107,7 @@ class Interface {
   void ProcessCommand();
   bool SendNextByte();
 
-  void SendResponse(ErrorCode error, uint32_t response_length);
+  void SendResponse(ErrorCode error, size_t response_length);
   void SendError(ErrorCode error) { SendResponse(error, 0); }
 };
